refactor: Structure fix
