+++ src/main/java/com/tenbit/beep/BeepApplication.java
@@ -1,11 +1,7 @@
 package com.tenbit.beep;
 
-import com.tenbit.beep.auth.domain.repository.UserRepository;
-import com.tenbit.beep.auth.domain.service.DBService;
-import com.tenbit.beep.auth.domain.service.impl.DBServiceImpl;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
 
 @SpringBootApplication
 public class BeepApplication {
