+++ src/main/java/com/tenbit/beep/attendance/service/impl/AttendanceServiceImpl.java
@@ -2,9 +2,9 @@ package com.tenbit.beep.attendance.service.impl;
 
 
 import com.tenbit.beep.attendance.service.AttendanceService;
-import com.tenbit.beep.auth.domain.domain.Attendance;
-import com.tenbit.beep.auth.domain.domain.User;
-import com.tenbit.beep.auth.domain.repository.UserRepository;
+import com.tenbit.beep.attendance.domain.Attendance;
+import com.tenbit.beep.auth.domain.User;
+import com.tenbit.beep.auth.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
