+++ src/main/java/com/tenbit/beep/auth/controller/RootController.java
@@ -1,0 +1,13 @@
+package com.tenbit.beep.auth.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class RootController {
+    
+    @GetMapping("/")
+    public String root() {
+        return "redirect:/main/main.html";
+    }
+}
