+++ src/main/java/com/tenbit/beep/auth/domain/Authority.java
@@ -1,0 +1,7 @@
+package com.tenbit.beep.auth.domain;
+
+public enum Authority {
+
+    TEACHER,
+    STUDENT
+}
