--- src/main/java/com/tenbit/beep/auth/domain/controller/RootController.java
@@ -1,13 +1,0 @@
-package com.tenbit.beep.auth.domain.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class RootController {
-    
-    @GetMapping("/")
-    public String root() {
-        return "redirect:/main/main.html";
-    }
-}
