--- src/main/java/com/tenbit/beep/auth/domain/domain/Attendance.java
@@ -1,9 +1,0 @@
-package com.tenbit.beep.auth.domain.domain;
-
-public enum Attendance {
-    TRUE,
-    FALSE
-}
-
-
-
