--- src/main/java/com/tenbit/beep/auth/domain/domain/Authority.java
@@ -1,7 +1,0 @@
-package com.tenbit.beep.auth.domain.domain;
-
-public enum Authority {
-
-    TEACHER,
-    STUDENT
-}
