--- src/main/java/com/tenbit/beep/auth/domain/dto/LoginRequest.java
@@ -1,12 +1,0 @@
-package com.tenbit.beep.auth.domain.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequest {
-
-    private String publicId;
-    private String password;
-}
