--- src/main/java/com/tenbit/beep/auth/domain/dto/SignupRequest.java
@@ -1,15 +1,0 @@
-package com.tenbit.beep.auth.domain.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SignupRequest {
-
-    private int studentNumber;
-//    private String name;
-    private String publicId;
-    private String password;
-    private String email;
-}
