--- src/main/java/com/tenbit/beep/auth/domain/exception/AuthenticationException.java
@@ -1,20 +1,0 @@
-package com.tenbit.beep.auth.domain.exception;
-
-public class AuthenticationException extends RuntimeException {
-
-    public AuthenticationException() {
-        super();
-    }
-
-    public AuthenticationException(String message) {
-        super(message);
-    }
-
-    public AuthenticationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public AuthenticationException(Throwable cause) {
-        super(cause);
-    }
-}
