--- src/main/java/com/tenbit/beep/auth/domain/exception/TokenExpiredException.java
@@ -1,20 +1,0 @@
-package com.tenbit.beep.auth.domain.exception;
-
-public class TokenExpiredException extends AuthenticationException {
-
-    public TokenExpiredException() {
-        super();
-    }
-
-    public TokenExpiredException(String message) {
-        super(message);
-    }
-
-    public TokenExpiredException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public TokenExpiredException(Throwable cause) {
-        super(cause);
-    }
-}
