--- src/main/java/com/tenbit/beep/auth/domain/exception/UserNotFoundException.java
@@ -1,20 +1,0 @@
-package com.tenbit.beep.auth.domain.exception;
-
-public class UserNotFoundException extends AuthenticationException {
-
-    public UserNotFoundException() {
-        super();
-    }
-
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-
-    public UserNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UserNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
