--- src/main/java/com/tenbit/beep/auth/domain/exception/ValueMissingException.java
@@ -1,20 +1,0 @@
-package com.tenbit.beep.auth.domain.exception;
-
-public class ValueMissingException extends AuthenticationException {
-
-    public ValueMissingException() {
-        super();
-    }
-
-    public ValueMissingException(String message) {
-        super(message);
-    }
-
-    public ValueMissingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ValueMissingException(Throwable cause) {
-        super(cause);
-    }
-}
