--- src/main/java/com/tenbit/beep/auth/domain/jwt/JwtUtil.java
@@ -1,16 +1,0 @@
-package com.tenbit.beep.auth.domain.jwt;
-
-import io.jsonwebtoken.security.Keys;
-
-import java.security.Key;
-
-public class JwtUtil {
-
-    protected final Key key;
-    protected final Long expiration;
-
-    public JwtUtil(String secretKey, Long expiration) {
-        this.key = Keys.hmacShaKeyFor(secretKey.getBytes());
-        this.expiration = expiration;
-    }
-}
