--- src/main/java/com/tenbit/beep/auth/domain/service/AuthService.java
@@ -1,10 +1,0 @@
-package com.tenbit.beep.auth.domain.service;
-
-import com.tenbit.beep.auth.domain.dto.LoginRequest;
-import com.tenbit.beep.auth.domain.dto.SignupRequest;
-
-public interface AuthService {
-
-    void signup(SignupRequest signupRequest);
-    String login(LoginRequest loginRequest);
-}
