--- src/main/java/com/tenbit/beep/auth/domain/service/DBService.java
@@ -1,9 +1,0 @@
-package com.tenbit.beep.auth.domain.service;
-
-public interface DBService {
-
-    void resetDB();
-//    void deleteByName(String name);
-    void deleteByPublicId(String publicId);
-    void deleteByInnerId(Long innerId);
-}
