--- src/main/java/com/tenbit/beep/auth/global/config/CorsConfig.java
@@ -1,24 +1,0 @@
-package com.tenbit.beep.auth.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig {
-
-    @Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                registry.addMapping("/**")
-                        .allowedOrigins("http://127.0.0.1:5500", "http://127.0.0.1:5500/auth/login", "http://localhost:5500", "http://localhost:5500/auth/login",
-                                "http://localhost:3000", "http://localhost:3000/signup")
-                        .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-                        .allowCredentials(true);
-            }
-        };
-    }
-}
