+++ src/main/java/com/tenbit/beep/auth/jwt/JwtTokenProvider.java
@@ -1,0 +1,25 @@
+package com.tenbit.beep.auth.jwt;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+
+import java.util.Date;
+
+public class JwtTokenProvider extends JwtUtil {
+
+    public JwtTokenProvider(String secretKey, Long expiration) {
+        super(secretKey, expiration);
+    }
+
+    public String generateToken(String publicId) {
+        Date now = new Date();
+        Date expiryDate = new Date(now.getTime() + expiration);
+
+        return Jwts.builder()
+                .setSubject(publicId)
+                .setIssuedAt(now)
+                .setExpiration(expiryDate)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+}
