+++ src/main/java/com/tenbit/beep/auth/jwt/JwtUtil.java
@@ -1,0 +1,16 @@
+package com.tenbit.beep.auth.jwt;
+
+import io.jsonwebtoken.security.Keys;
+
+import java.security.Key;
+
+public class JwtUtil {
+
+    protected final Key key;
+    protected final Long expiration;
+
+    public JwtUtil(String secretKey, Long expiration) {
+        this.key = Keys.hmacShaKeyFor(secretKey.getBytes());
+        this.expiration = expiration;
+    }
+}
