+++ src/main/java/com/tenbit/beep/auth/service/DBService.java
@@ -1,0 +1,9 @@
+package com.tenbit.beep.auth.service;
+
+public interface DBService {
+
+    void resetDB();
+//    void deleteByName(String name);
+    void deleteByPublicId(String publicId);
+    void deleteByInnerId(Long innerId);
+}
