+++ src/main/java/com/tenbit/beep/common/exception/AuthenticationException.java
@@ -1,0 +1,20 @@
+package com.tenbit.beep.common.exception;
+
+public class AuthenticationException extends RuntimeException {
+
+    public AuthenticationException() {
+        super();
+    }
+
+    public AuthenticationException(String message) {
+        super(message);
+    }
+
+    public AuthenticationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public AuthenticationException(Throwable cause) {
+        super(cause);
+    }
+}
