+++ src/main/java/com/tenbit/beep/common/exception/IllegalArgumentsException.java
@@ -1,0 +1,20 @@
+package com.tenbit.beep.common.exception;
+
+public class IllegalArgumentsException extends AuthenticationException {
+
+    public IllegalArgumentsException() {
+        super();
+    }
+
+    public IllegalArgumentsException(String message) {
+        super(message);
+    }
+
+    public IllegalArgumentsException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public IllegalArgumentsException(Throwable cause) {
+        super(cause);
+    }
+}
