+++ src/main/java/com/tenbit/beep/common/exception/TokenExpiredException.java
@@ -1,0 +1,20 @@
+package com.tenbit.beep.common.exception;
+
+public class TokenExpiredException extends AuthenticationException {
+
+    public TokenExpiredException() {
+        super();
+    }
+
+    public TokenExpiredException(String message) {
+        super(message);
+    }
+
+    public TokenExpiredException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public TokenExpiredException(Throwable cause) {
+        super(cause);
+    }
+}
