+++ src/main/java/com/tenbit/beep/common/exception/UserNotFoundException.java
@@ -1,0 +1,20 @@
+package com.tenbit.beep.common.exception;
+
+public class UserNotFoundException extends AuthenticationException {
+
+    public UserNotFoundException() {
+        super();
+    }
+
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+
+    public UserNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UserNotFoundException(Throwable cause) {
+        super(cause);
+    }
+}
